requirements.txt
# Core dependencies
python-telegram-bot==20.7
openai==1.3.7
anthropic==0.18.1
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1

# Caching and async
redis==5.0.1
aioredis==2.0.1

# HTTP requests
requests==2.31.0
aiohttp==3.9.1

# Logging and utilities
colorama==0.4.6
structlog==23.2.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Security
bandit==1.7.5
# safety==2.3.5  # Commented out due to dependency conflicts

# Optional: PostgreSQL support
psycopg2-binary==2.9.9 